# spring_ajax_jpa/openapi.yaml
openapi: 3.0.1
info:
  title: Game Management Service
  description: API for managing computer games, developers and publishers
  version: 1.0.0
servers:
  - url: http://localhost:8080/ComputerGames
...
paths:
  /games:
    get:
      tags: [Games]
      summary: Get all games
      responses:
        '200':
          description: List of all games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
                  
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'

    post:
      tags: [Games]
      summary: Create a new game
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Game title
                  example: "The Witcher 3: Wild Hunt"
                developer:
                  type: integer
                  description: Game developer identifier
                  example: 1
                publisher:
                  type: integer
                  description: Game publisher identifier
                  example: 1
                price:
                  type: number
                  format: double
                  description: Game price
                  example: 39.99
                releaseDate:
                  type: string
                  format: date
                  description: Game release date
                  example: 2015-05-19
              required:
                - title
                - developer
                - publisher
                - price
                - releaseDate
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
  /games/{id}:
    delete:
      tags: [Games]
      summary: Delete a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Game deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
    get:
      tags: [Games]
      summary: Get a game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Game finded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
        
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
...
components:
  schemas:
    Game:
      type: object
      properties:
        id:
          type: integer
          description: Game identifier
          example: 1
        title:
          type: string
          description: Game title
          example: "The Witcher 3: Wild Hunt"
        price:
          type: number
          description: Game price
          format: double
          example: 39.99
        releaseDate:
          type: string
          format: date
          description: Game release data
          example: 2015-05-19
        developer:
          type: integer
          description: Game developer identifier
          example: 1
        publisher:
          type: integer
          description: Game publisher identifier
          example: 1
      required:
        - id
        - title
        - price
        - releaseDate
        - developer
        - publisher
...
    OperationStatus:
      type: object
      properties:
        status:
          type: string
          description: Description of the operation execution status
          example: Operation error, invalid fields